Updated the reference table and pricing details
Usage of NIC in the reference table could lead to a mis-understanding that multiple-IP addresses are possible for a single NIC.
IPs on a ARM IaaS VM is similar to an ILPIP in classic model. The first IP address on a VM is being charged currently. This will change shortly in future.